Fix navigating back with Turbolinks
When navigating back, Turbolinks does not properly restore pages if
`history.replaceState` was called with a non-Turbolinks state object.
This is probably a bug.  The workaround is to use a state object that
has a `turbolinks` property, which `history.state` does if Turbolinks is
active (thus allowing moar.js to remain Turbolinks-agnostic).

See https://github.com/turbolinks/turbolinks/issues/219